requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
idna==3.3
libretranslatepy==2.1.1
lxml==4.8.0
oauthlib==3.2.0
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
translate==3.6.1
tweepy==4.8.0
urllib3==1.26.9
